% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dep.R
\name{check_dep}
\alias{check_dep}
\title{Check attached dependency}
\usage{
check_dep(pkg)
}
\arguments{
\item{pkg}{a character string of the package name}
}
\value{
Null
}
\description{
Stop execution if a package is not attached.
}
\keyword{internal}
