% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicates.R
\name{check_duplicates}
\alias{check_duplicates}
\title{Check for duplicates}
\usage{
check_duplicates(x)
}
\arguments{
\item{x}{A vector or list.}
}
\description{
Checks for duplicated items in a vector or list and throw an error if found.
}
\keyword{internal}
