% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_genome_build.R
\name{check_genome_build}
\alias{check_genome_build}
\title{Check genome build}
\usage{
check_genome_build(genome_build)
}
\arguments{
\item{genome_build}{A character string with the abbreviated genome build
name, or a BSGenome object. At the moment, only \code{hg38} and \code{hg19}
are supported as abbreviated input.}
}
\value{
A BSGenome object.
}
\description{
Check if the genome build is valid and return the appropriate BSGenome
object.
}
\examples{
if (requireNamespace("BSgenome.Hsapiens.UCSC.hg38", quietly = TRUE)) {
    check_genome_build("hg38")
}

}
\seealso{
\link[BSgenome]{BSgenome-class} for more information on BSGenome
objects.
}
