% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CTCF_ChIP_peaks}
\alias{CTCF_ChIP_peaks}
\title{Example ChIP-seq peak file}
\format{
An object of class \code{GRanges} of length 209.
}
\source{
\href{https://www.encodeproject.org/files/ENCFF091ODJ/}{ENCODE Accession: ENCFF091ODJ}
}
\usage{
data("CTCF_ChIP_peaks")
}
\description{
Human CTCF peak file generated with ChIP-seq using HCT116 cell-line. No
control files were used to generate the peak file.
}
\note{
To reduce the size of the package, the included peak file focuses on
specific genomic regions. The subset region included is
chr10:65,654,529-74,841,155.
}
\keyword{datasets}
