% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CTCF_TIP_peaks}
\alias{CTCF_TIP_peaks}
\title{Example TIP-seq peak file}
\format{
An object of class \code{GRanges} of length 182.
}
\usage{
data("CTCF_TIP_peaks")
}
\description{
Human CTCF peak file generated with TIP-seq using HCT116 cell-line.
The peak file was generated using the
\href{https://nf-co.re/cutandrun/3.2.2}{nf-core/cutandrun} pipeline.
Raw read files were sourced from \emph{NIH Sequence Read Archives}
(ID: \href{https://trace.ncbi.nlm.nih.gov/Traces/index.html?view=run_browser&acc=SRR16963166}{SRR16963166}).
}
\note{
To reduce the size of the package, the included peak file focuses on
specific genomic regions. The subset region included is
chr10:65,654,529-74,841,155.
}
\keyword{datasets}
