% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{format_exptype}
\alias{format_exptype}
\title{Format exp_type}
\usage{
format_exptype(exp_type)
}
\arguments{
\item{exp_type}{A character depicting the type of experiment.
Supported experimental types are:
\itemize{
  \item \code{chipseq}: ChIP-seq data
  \item \code{tipseq}: TIP-seq data
  \item \code{cuttag}: CUT&Tag data
  \item \code{cutrun}: CUT&Run data
  \item \code{other}: Other experiment type data
  \item \code{unknown}: Unknown experiment type data
}
Any item not mentioned above will be returned as-is.}
}
\value{
A character vector of formatted exp_type.
}
\description{
Format input exp_type to look pretty.
}
\examples{
MotifPeeker:::format_exptype("chipseq")

}
\keyword{internal}
