% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bpparam.R
\name{get_bpparam}
\alias{get_bpparam}
\title{Get parameters for \link[BiocParallel]{BiocParallel}}
\usage{
get_bpparam(
  workers,
  progressbar = workers > 1,
  force_snowparam = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{workers}{The number of workers to use for parallel processing.}

\item{progressbar}{
    \code{logical(1)} Enable progress bar (based on plyr:::progress_text).
  }

\item{force_snowparam}{A logical indicating whether to force the use of
\link[BiocParallel]{SnowParam} object.}

\item{verbose}{A logical indicating whether to print verbose messages while
running the function. (default = FALSE)}
}
\value{
A \code{BPPARAM} object.
}
\description{
Get appropriate parameters for \code{BiocParallel} based on the
number of workers specified. For less than 10 workers, the function returns a
\code{MulticoreParam} object. For 10 or more cores, the function
returns a \code{SnowParam} object. Since Windows supports
neither, the function returns a \code{SerialParam} object. As a
result, Windows users do not benefit from parallel processing.
}
\seealso{
\link[BiocParallel]{BiocParallelParam}
}
\keyword{internal}
