% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_download_buttons.R
\name{get_download_buttons}
\alias{get_download_buttons}
\title{Get download buttons for peak file, STREME and TOMOTM output}
\usage{
get_download_buttons(
  comparison_i,
  start_i,
  segregated_peaks,
  out_dir,
  add_buttons = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{comparison_i}{Index of the comparison pair group.}

\item{start_i}{Index of the first comparison pair.}

\item{segregated_peaks}{A list of peak files generated from
\code{\link{segregate_seqs}}.}

\item{out_dir}{A character vector of the directory with STREME and TOMTOM
output.}

\item{add_buttons}{A logical indicating whether to prepare download buttons.}

\item{verbose}{A logical indicating whether to print messages.}
}
\value{
A list of download buttons for peak file, STREME and TOMTOM output.
}
\description{
Get download buttons for peak file, STREME and TOMOTM output
}
\keyword{internal}
