% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_JASPARCORE.R
\name{get_JASPARCORE}
\alias{get_JASPARCORE}
\title{Download JASPAR CORE database}
\usage{
get_JASPARCORE(verbose = FALSE)
}
\arguments{
\item{verbose}{A logical indicating whether to print verbose messages while
running the function. (default = FALSE)}
}
\value{
A character string specifying the path to the downloaded file
(\code{meme} format).
}
\description{
Downloads JASPAR CORE database in \code{meme} format for all available
taxonomic groups. Uses BiocFileCache to cache downloads.
}
\examples{
get_JASPARCORE()

}
