% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_JASPAR.R
\name{link_JASPAR}
\alias{link_JASPAR}
\title{Get JASPAR link for motifs}
\usage{
link_JASPAR(motif_id, download = FALSE)
}
\arguments{
\item{motif_id}{A character string specifying the JASPAR motif ID.}

\item{download}{A logical specifying whether to return a download link or an
HTML embeddable matrix profile link. (default = FALSE)}
}
\value{
A character string containing the JASPAR motif link.
}
\description{
Get JASPAR link for motifs
}
\keyword{internal}
