% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{messager}
\alias{messager}
\title{Print messages}
\usage{
messager(..., v = Sys.getenv("VERBOSE") != "FALSE")
}
\arguments{
\item{v}{Whether to print messages or not.}
}
\value{
Null
}
\description{
Conditionally print messages.
 Allows developers to easily control verbosity of functions, 
 and meet Bioconductor requirements that dictate the message 
 must first be stored to a variable before passing to \link[base]{message}.
}
\keyword{internal}
