% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motif_MA1102.3}
\alias{motif_MA1102.3}
\title{Example CTCFL JASPAR motif file}
\format{
An object of class \code{universalmotif} of length 1.
}
\source{
\href{https://jaspar.elixir.no/matrix/MA1102.3/}{JASPAR Matrix ID: MA1102.3}
}
\usage{
data("motif_MA1102.3")
}
\description{
The motif file contains the JASPAR motif for CTCFL (MA1102.3) for
\emph{Homo Sapiens}. This is one of the two motif files used to demonstrate
\code{\link{MotifPeeker}}'s known-motif analysis functionality.
}
\keyword{datasets}
