% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motif_MA1930.2}
\alias{motif_MA1930.2}
\title{Example CTCF JASPAR motif file}
\format{
An object of class \code{universalmotif} of length 1.
}
\source{
\href{https://jaspar.elixir.no/matrix/MA1930.2/}{JASPAR Matrix ID: MA1930.2}
}
\usage{
data("motif_MA1930.2")
}
\description{
The motif file contains the JASPAR motif for CTCF (MA1930.2) for
\emph{Homo Sapiens}. This is one of the two motif files used to demonstrate
\code{\link{MotifPeeker}}'s known-motif analysis functionality.
}
\keyword{datasets}
