% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifPeeker.R
\name{MotifPeeker}
\alias{MotifPeeker}
\title{Benchmark epigenomic profiling methods using motif enrichment}
\usage{
MotifPeeker(
  peak_files = NULL,
  alignment_files = NULL,
  control_files = NULL,
  labels = NULL,
  exp_type = NULL,
  genome_build,
  motif_files = NULL,
  motif_labels = NULL,
  reference_index = 1,
  cell_counts = NULL,
  peakcalling_qvalue = 0.01,
  denovo_motif_discovery = TRUE,
  denovo_motifs = 3,
  download_buttons = TRUE,
  input_dir = ".",
  output_dir = tempdir(),
  overwrite = FALSE,
  use_cache = TRUE,
  ncpus = 1,
  quiet = FALSE,
  debug = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{labels}{A character vector of labels for each peak file.}

\item{peakfiles}{A character vector of peak file names in narrowPeak

(MACS2/3 output) or BED (SEACR) format.}
}
\description{
This function compares different epigenomic datasets using motif enrichment
as the key metric. The output is an easy-to-interpret HTML document with the
results. The report contains three main sections: (1) General Metrics on peak
and alignment files (if provided), (2) Known Motif Enrichment Analysis and
(3) De-novo Motif Enrichment Analysis.
}
\details{
The function can optionally call peaks using MACS3, but it is recommended
that the peak files be supplied directly and produced using fine-tuned
parameters.
}
