% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_paths.R
\name{normalise_paths}
\alias{normalise_paths}
\title{Apply \code{\link[base]{normalizePath}} to a list of paths}
\usage{
normalise_paths(path_list)
}
\arguments{
\item{path_list}{A list of paths.}
}
\value{
A list of normalised paths or the input as is if contents are not
a character.
}
\description{
Apply \code{\link[base]{normalizePath}} to a list of paths
}
\keyword{internal}
