% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_DT.R
\name{print_DT}
\alias{print_DT}
\title{Print DT table}
\usage{
print_DT(df, html_tags = FALSE, extra = FALSE)
}
\arguments{
\item{df}{Dataframe/tibble to be printed.}

\item{html_tags}{Logical. If TRUE, returns the table as a tagList object.}

\item{extra}{Logical. If TRUE, adds extra options like search to the
datatable.}
}
\value{
A DT object suitable to be used with \code{print()}.
}
\description{
Print DT table
}
\keyword{internal}
