% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{random_string}
\alias{random_string}
\title{Generate a random string}
\usage{
random_string(length)
}
\arguments{
\item{length}{The length of the random string to generate.}
}
\value{
A random string of the specified length.
}
\description{
Generate a random string
}
\keyword{internal}
