% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_command.R
\name{report_command}
\alias{report_command}
\title{Report command}
\usage{
report_command(params)
}
\arguments{
\item{params}{A list of parameters used to generate the HTML report.}
}
\value{
A character string containing the reconstructed
\code{\link{MotifPeeker}} command.
}
\description{
Reconstruct the \code{\link{MotifPeeker}} command from the parameters used to
generate the HTML report.
}
\examples{
report_command(params = list(
   alignment_files = c("file1.bam", "file2.bam"),
   exp_labels = c("exp1", "exp2"),
   genome_build = "hg19"))


}
