% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_header.R
\name{report_header}
\alias{report_header}
\title{Report header}
\usage{
report_header()
}
\value{
Header string to be rendering within Rmarkdown file.
}
\description{
Credit: This function was adapted from the \emph{EpiCompare} package.
}
\details{
Generate a header for \link[MotifPeeker]{MotifPeeker} reports generated using 
the \emph{MotifPeeker.Rmd} template.
}
\examples{
report_header()

}
\keyword{internal}
