% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_cache.R
\name{use_cache}
\alias{use_cache}
\title{Check, add and access files in cache}
\usage{
use_cache(url, verbose = FALSE)
}
\arguments{
\item{url}{A character string specifying the URL of the file to check for.}

\item{verbose}{A logical indicating whether to print verbose messages while
running the function. (default = FALSE)}
}
\value{
A character string specifying the path to the cached file.
}
\description{
Query local BiocFileCache to get cached version of a file and add them if
they do not exist.
}
\keyword{internal}
