% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dep.R
\name{check_dep}
\alias{check_dep}
\title{Check attached dependency}
\usage{
check_dep(pkg, fatal = TRUE, custom_msg = NULL)
}
\arguments{
\item{pkg}{a character string of the package name}

\item{fatal}{a logical value indicating whether to stop execution if the
package is not attached.}

\item{custom_msg}{a custom message to display if the package is not attached.}
}
\value{
Null
}
\description{
Stop execution if a package is not attached.
}
\keyword{internal}
