% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pretty_number}
\alias{pretty_number}
\title{Convert numbers to more readable strings}
\usage{
pretty_number(x, decimal_digits = 2)
}
\arguments{
\item{x}{A number.}

\item{decimal_digits}{Number of decimal digits to round to.}
}
\value{
A character string of the formatted number. \code{NA} is returned as
"NA".
}
\description{
Format raw numbers to more readable strings. For example, 1000 will be
converted to "1K". Supported suffixes are "K", "M", and "B".
}
\examples{
print(MotifPeeker:::pretty_number(134999))

}
\keyword{internal}
